Added docs for jsonIter Refactored code added option def to jsonIterApi
